requirements.txt
# OpenAI
openai>=1.0.0
python-dotenv>=1.0.0

# Vector DB
chromadb>=0.4.0

# UI
streamlit>=1.32.0
